[IOTDB-1507]Move MLogTxtWriter to logfile package
